requirements.txt
opencv-python==4.2.0.32
tqdm
Pillow
scipy
albumentations
pandas
